requirements.txt
azure-identity==1.15.0
azure-cosmos==4.5.1
Babel==2.14.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ghp-import==2.1.0
idna==3.6
importlib-metadata==7.0.0
Jinja2==3.1.2
Markdown==3.5.1
MarkupSafe==2.1.3
mergedeep==1.3.4
mkdocs==1.5.3
mkdocs-material==9.5.2
mkdocs-material-extensions==1.3.1
packaging==23.2
paginate==0.5.6
pathspec==0.12.1
platformdirs==4.1.0
Pygments==2.17.2
pymdown-extensions==10.5
python-dateutil==2.8.2
PyYAML==6.0.1
pyyaml_env_tag==0.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
urllib3==2.1.0
watchdog==3.0.0
zipp==3.17.0
